Added suspend basal & bolus cancel tests. Seperated Bolus tests from Message tests.
