+++ OmniKitTests/BasalScheduleTests.swift
@@ -119,5 +119,21 @@ class BasalScheduleTests: XCTestCase {
         XCTAssertEqual(0.05, rateEntry.rate)
         XCTAssertEqual(TimeInterval(hours: 24), rateEntry.duration, accuracy: 0.001)
     }
+    
+    func testSuspendBasalCommand() {
+        do {
+            // Decode 1f 05 6fede14a 01
+            let cmd = try CancelDeliveryCommand(encodedData: Data(hexadecimalString: "1f056fede14a01")!)
+            XCTAssertEqual(0x6fede14a, cmd.nonce)
+            XCTAssertEqual(.noBeep, cmd.beepType)
+            XCTAssertEqual(.basal, cmd.deliveryType)
+        } catch (let error) {
+            XCTFail("message decoding threw error: \(error)")
+        }
+        
+        // Encode
+        let cmd = CancelDeliveryCommand(nonce: 0x6fede14a, deliveryType: .basal, beepType: .noBeep)
+        XCTAssertEqual("1f056fede14a01", cmd.data.hexadecimalString)
+    }
 }
